% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_backup.R
\name{pkg_backup}
\alias{pkg_backup}
\title{Package Backup}
\usage{
pkg_backup(lib.loc = NULL)
}
\arguments{
\item{lib.loc}{Library Location. It's an argument passed to
\code{installed.packages()}.}
}
\description{
\code{pkg_backup()} saves the names of installed packages to a temporary
online drive and provides a download code if the process is successful.
The download code can be used in \code{pkg_restore()} to re-install packages
on a different version of R or a different computer.
}

