% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_copy.R
\name{pkg_copy}
\alias{pkg_copy}
\title{Copy package info to cloud}
\usage{
pkg_copy(lib.loc = NULL)
}
\arguments{
\item{lib.loc}{Library Location. It's an argument passed to
\code{installed.packages()}.}
}
\description{
\code{pkg_copy()} saves the names of local R packages to a temporary
cloud drive and provides a download code if the process is successful.
The download code can be used in \code{pkg_paste()} to re-install packages
in a different environment. \cr\cr This function should be run in the
old/source environment.
}
\examples{
pkg_copy()

}

