% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_paste.R
\name{pkg_paste}
\alias{pkg_paste}
\alias{pkg_paste_names}
\title{Paste copied packages to another environment}
\usage{
pkg_paste(code, lib.loc = NULL, ...)

pkg_paste_names(code, lib.loc = NULL)
}
\arguments{
\item{code}{Download code returned by \code{pkg_backup()}.}

\item{lib.loc}{Library Location. It's an argument passed to
\code{installed.packages()}.}

\item{...}{Arguments passed to \code{install.packages()}}
}
\description{
\code{pkg_paste()} retrieve copied R package information through the
download code generated by \code{pkg_copy()}. It then compares it with the
package information in the current environment. If it detects any new
packages, it will attempt to install them and report any packages that
cannot be installed. \code{pkg_paste_names()} does not attempt to install
packages, in case people want to revise the list before installation. \cr\cr
This function should be run in the new/target environment.
}

