% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_restore.R
\name{pkg_restore}
\alias{pkg_restore}
\alias{pkg_restore_list}
\title{Package Restore}
\usage{
pkg_restore(code, lib.loc = NULL, ...)

pkg_restore_list(code, lib.loc = NULL)
}
\arguments{
\item{code}{Download code returned by \code{pkg_backup()}.}

\item{lib.loc}{Library Location. It's an argument passed to
\code{installed.packages()}.}

\item{...}{Arguments passed to \code{install.packages()}}
}
\description{
\code{pkg_restore()} obtains the list of saved R packages and makes a
comparison with the list of available packages on the new machine (or new R
environment). It will attempt to install all the packages that are not
available on the new machine and report any packages that cannot be
installed. \code{pkg_restore_list()} only provides package names and does
not attempt to install. Users can pass this list to \code{install.packages()}
to install by themselves.
}

